pyproject.toml
[project]
name = "solar-deep-research"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "jupyter>=1.1.1",
    "langchain-community>=0.3.26",
    "langchain-core>=0.3.67",
    "langchain-ollama>=0.3.8",
    "langchain-openai>=0.3.27",
    "langchain-qdrant>=0.2.0",
    "langgraph>=0.5.0",
    "langsmith>=0.4.4",
    "pandas>=2.3.0",
    "pdfminer.six>=20221105",
    "pymupdf>=1.26.1",
    "pymupdf4llm>=0.0.27",
    "unstructured[pdf]>=0.18.15",
    "unstructured-inference>=0.7.0",
    "pi_heif>=0.16.0",
    "pdf2image>=1.17.0",
    "pytesseract>=0.3.13",
    "docling>=2.55.1",
    "playwright>=1.55.0",
    "pdfplumber>=0.11.7",
    "tabula-py>=2.10.0",
    "geopy>=2.4.1",
    "uvicorn>=0.38.0",
    "fastapi>=0.119.1",
    "ragas>=0.3.7",
]
